requirements.txt
# Inteligencia de Datos SUDEASEG - Sector Asegurador Venezolano
# Python 3.10+

# Web scraping y red
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Extracción PDF (Camelot requiere ghostscript en el sistema)
camelot-py[cv]>=0.11.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
openpyxl>=3.1.0
xlrd>=2.0.0

# ETL y datos
pandas>=2.1.0
numpy>=1.26.0

# Tasas de cambio BCV / APIs
# pyBCV o requests para bcv.org.ve / DolarApi

# Base de datos y backend
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0
altair>=5.2.0

# OCR para PDFs escaneados (anuarios antiguos)
pytesseract>=0.3.10
Pillow>=10.0.0

# Utilidades
python-dotenv>=1.0.0
schedule>=1.2.0
tenacity>=8.2.0
